Testing out SIMD

+++ benches/search_trees.rs
@@ -4,7 +4,7 @@ use rand::random;
 use obliviousdb::search_tree::SearchTree;
 
 fn benchmark_search_oblivious_static_search_tree(c: &mut Criterion) {
-    let max: i64 = 268_435_456_0;/*536_870_912_0;*//*1_073_741_824_0;*/
+    let max: u32 = 268_435_456_0;/*536_870_912_0;*//*1_073_741_824_0;*/
     let min = 0;
     let gen = (min..max).step_by(10).into_iter();
 
@@ -19,16 +19,16 @@ fn benchmark_search_oblivious_static_search_tree(c: &mut Criterion) {
 
 fn benchmark_search_std_collection_btreemap(c: &mut Criterion) {
     use std::collections::BTreeMap;
-    let max: i64 = 268_435_456_0;/*536_870_912_0;*//*1_073_741_824_0;*/
+    let max: u32 = 268_435_456_0;/*536_870_912_0;*//*1_073_741_824_0;*/
     let min = 0;
     let gen = (min..max).step_by(10).into_iter();
 
-    let mut tree = BTreeMap::<i64, bool>::new();
+    let mut tree = BTreeMap::<u32, bool>::new();
     gen.for_each(|i| { tree.insert(i, true); });
 
     c.bench_function("btreemap search",
     |b| b.iter(|| {
-        let element: i64 = random();//i.next().unwrap();
+        let element: u32 = random();//i.next().unwrap();
         tree.contains_key(&element)
     }));
 }
