+++ src/search_tree/util.rs
@@ -2,12 +2,12 @@ pub fn is_odd(element: u16) -> bool {
     return (1 & element) > 0
 }
 
-pub fn size_of_tree_with_height(height: u16) -> u64 {
+pub fn size_of_tree_with_height(height: u16) -> u32 {
     // 2 ^ (height) - 1
     (1 << height) - 1
 }
 
-pub fn number_of_leaves_in_tree(height: u16) -> u64 {
+pub fn number_of_leaves_in_tree(height: u16) -> u32 {
     // 2^(height-1)
     1 << (height - 1)
 }
